test_requirements.txt
tox==3.17.0
factory-boy==2.12.0
pytest==4.6
pytest-cov==2.11.0
pytest-django==3.2.1
pytz==2020.1
mock==3.0.5
django-model-utils==2.3.1
python-dateutil==2.8.1
flake8==3.8.3
django-waffle==0.18.0

# XBlock dependencies to make testing easier
# mostly to import workbench.settings
xblock-utils>=2.1.1
XBlock==1.2.1; python_version < '3'
XBlock>=1.4.0; python_version >= '3'
xblock-sdk>=0.2.2
django-pyfs>=2.2
lazy>=1.4
